+++ riverpod_for_serverpod_runtime/lib/riverpod_for_serverpod_runtime.dart
@@ -4,4 +4,7 @@ export 'src/cache/cached_entity_record.dart';
 export 'src/cache/cached_index_record.dart';
 export 'src/cache/generated_cache_storage.dart';
 export 'src/cache/generated_entity_cache.dart';
+export 'src/cache/generated_key_value_storage.dart';
+export 'src/cache/json_generated_cache_storage.dart';
 export 'src/cache/memory_generated_cache_storage.dart';
+export 'src/cache/memory_generated_key_value_storage.dart';